fastapi-ai-agent/requirements.txt
FastAPI
uvicorn
transformers
torch
git+https://github.com/superannotateai/generated_text_detector.git@v1.1.0 
langchain-openai
langchain-core
python-dotenv
requests
pydantic
langchain-community
langchain_google_community
ultralytics
pillow
inference_sdk
efficientnet_pytorch
torchvision
git+https://github.com/albu/albumentations@529b074309256272fd2f0b2fdd52ccc62183fc57
